testing display of multiple editors in browser at same time
