+++ src/bundler/plugins/fetch-plugin.ts
