+++ src/index.tsx
@@ -1,32 +1,12 @@
 import 'bulmaswatch/superhero/bulmaswatch.min.css';
-
-import { useState, useEffect, useRef } from 'react';
 import ReactDOM from 'react-dom';
-
-import CodeEditor from './components/code-editor';
-import Preview from './components/preview';
-import bundle from './bundler';
+import CodeCell from './components/code-cell';
 
 const App = () => {
-  const [code, setCode] = useState('');
-  const [input, setInput] = useState('');
-
-  const onClick = async () => {
-    const output = await bundle(input);
-    setCode(output);
-  };
-
   return (
     <div>
-      {/* initial value of editor is passed below */}
-      <CodeEditor
-        initialValue="const a = 1;"
-        onChange={(value) => setInput(value)}
-      />
-      <div>
-        <button onClick={onClick}>Submit</button>
-      </div>
-      <Preview code={code} />
+      <CodeCell />
+      <CodeCell />
     </div>
   );
 };
